requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
autopep8==1.6.0
boto3==1.20.41
botocore==1.23.41
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptowatch-sdk==0.0.16
dataclasses==0.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.0.2
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==2.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
MarkupSafe==2.0.1
marshmallow==3.13.0
multidict==6.1.0
nltk==3.9.1
numerize==0.12
numpy==1.26.0
openai==0.27.7
# pandas==2.1.0
pandas==1.3.5
propcache==0.2.1
protobuf==3.19.3
pycodestyle==2.8.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.25.1
s3transfer==0.5.0
six==1.16.0
sniffio==1.3.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.2
yarl==1.18.3
zipp==3.6.0
